Revert "New Feature : Allow inviting new users from the card assignment dialog"
