+++ src/components/InteractiveStepWrapper.tsx
@@ -56,11 +56,6 @@ type InteractiveStepWrapperProps = {
      * This flag can be removed, once all components/screens have switched to edge-to-edge safe area handling.
      */
     enableEdgeToEdgeBottomSafeAreaPadding?: boolean;
-
-    /**
-     * Callback to be called when the screen entry transition ends.
-     */
-    onEntryTransitionEnd?: () => void;
 };
 
 function InteractiveStepWrapper(
@@ -79,7 +74,6 @@ function InteractiveStepWrapper(
         offlineIndicatorStyle,
         shouldKeyboardOffsetBottomSafeAreaPadding,
         enableEdgeToEdgeBottomSafeAreaPadding,
-        onEntryTransitionEnd,
     }: InteractiveStepWrapperProps,
     ref: React.ForwardedRef<View>,
 ) {
@@ -97,7 +91,6 @@ function InteractiveStepWrapper(
             shouldShowOfflineIndicatorInWideScreen={shouldShowOfflineIndicatorInWideScreen}
             offlineIndicatorStyle={offlineIndicatorStyle}
             shouldKeyboardOffsetBottomSafeAreaPadding={shouldKeyboardOffsetBottomSafeAreaPadding}
-            onEntryTransitionEnd={onEntryTransitionEnd}
         >
             <HeaderWithBackButton
                 title={headerTitle}
