+++ src/pages/workspace/companyCards/assignCard/AssignCardFeedPage.tsx
@@ -19,7 +19,6 @@ import AssigneeStep from './AssigneeStep';
 import CardNameStep from './CardNameStep';
 import CardSelectionStep from './CardSelectionStep';
 import ConfirmationStep from './ConfirmationStep';
-import InviteNewMemberStep from './InviteNewMemberStep';
 import TransactionStartDateStep from './TransactionStartDateStep';
 
 type AssignCardFeedPageProps = PlatformStackScreenProps<SettingsNavigatorParamList, typeof SCREENS.WORKSPACE.COMPANY_CARDS_ASSIGN_CARD> & WithPolicyAndFullscreenLoadingProps;
@@ -92,13 +91,6 @@ function AssignCardFeedPage({route, policy}: AssignCardFeedPageProps) {
                     policyID={policyID}
                     feed={feed}
                     backTo={shouldUseBackToParam ? backTo : undefined}
-                />
-            );
-        case CONST.COMPANY_CARD.STEP.INVITE_NEW_MEMBER:
-            return (
-                <InviteNewMemberStep
-                    route={route}
-                    feed={feed}
                 />
             );
         default:
