+++ src/pages/workspace/companyCards/assignCard/AssigneeStep.tsx
@@ -1,6 +1,7 @@
-import React, {useEffect, useMemo, useState} from 'react';
+import React, {useMemo, useState} from 'react';
 import {Keyboard} from 'react-native';
 import type {OnyxEntry} from 'react-native-onyx';
+import FormAlertWithSubmitButton from '@components/FormAlertWithSubmitButton';
 import * as Expensicons from '@components/Icon/Expensicons';
 import InteractiveStepWrapper from '@components/InteractiveStepWrapper';
 import SelectionList from '@components/SelectionListWithSections';
@@ -9,17 +10,15 @@ import UserListItem from '@components/SelectionListWithSections/UserListItem';
 import Text from '@components/Text';
 import useCardFeeds from '@hooks/useCardFeeds';
 import useCardsList from '@hooks/useCardsList';
+import useDebouncedState from '@hooks/useDebouncedState';
 import useLocalize from '@hooks/useLocalize';
 import useNetwork from '@hooks/useNetwork';
 import useOnyx from '@hooks/useOnyx';
-import useSearchSelector from '@hooks/useSearchSelector';
 import useThemeStyles from '@hooks/useThemeStyles';
-import {setDraftInviteAccountID} from '@libs/actions/Card';
-import {searchInServer} from '@libs/actions/Report';
 import {getDefaultCardName, getFilteredCardList, hasOnlyOneCardToAssign} from '@libs/CardUtils';
 import {getHeaderMessage, getSearchValueForPhoneOrEmail, sortAlphabetically} from '@libs/OptionsListUtils';
 import {getPersonalDetailByEmail} from '@libs/PersonalDetailsUtils';
-import {getIneligibleInvitees, isDeletedPolicyEmployee} from '@libs/PolicyUtils';
+import {isDeletedPolicyEmployee} from '@libs/PolicyUtils';
 import tokenizedSearch from '@libs/tokenizedSearch';
 import Navigation from '@navigation/Navigation';
 import {setAssignCardStepAndData} from '@userActions/CompanyCards';
@@ -27,6 +26,8 @@ import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type * as OnyxTypes from '@src/types/onyx';
 import type {AssignCardData, AssignCardStep} from '@src/types/onyx/AssignCard';
+
+const MINIMUM_MEMBER_TO_SHOW_SEARCH = 8;
 
 type AssigneeStepProps = {
     /** The policy that the card will be issued under */
@@ -46,42 +47,22 @@ function AssigneeStep({policy, feed}: AssigneeStepProps) {
     const [list] = useCardsList(policy?.id, feed);
     const [cardFeeds] = useCardFeeds(policy?.id);
     const filteredCardList = getFilteredCardList(list, cardFeeds?.settings?.oAuthAccountDetails?.[feed], workspaceCardFeeds);
-    const [didScreenTransitionEnd, setDidScreenTransitionEnd] = useState(false);
-    const [isSearchingForReports] = useOnyx(ONYXKEYS.IS_SEARCHING_FOR_REPORTS, {initWithStoredValues: false, canBeMissing: true});
-
-    const excludedUsers = useMemo(() => {
-        const ineligibleInvites = getIneligibleInvitees(policy?.employeeList);
-        return ineligibleInvites.reduce(
-            (acc, login) => {
-                acc[login] = true;
-                return acc;
-            },
-            {} as Record<string, boolean>,
-        );
-    }, [policy?.employeeList]);
-
-    const {searchTerm, setSearchTerm, debouncedSearchTerm, availableOptions, selectedOptionsForDisplay, areOptionsInitialized} = useSearchSelector({
-        selectionMode: CONST.SEARCH_SELECTOR.SELECTION_MODE_MULTI,
-        searchContext: CONST.SEARCH_SELECTOR.SEARCH_CONTEXT_MEMBER_INVITE,
-        includeUserToInvite: true,
-        excludeLogins: excludedUsers,
-        includeRecentReports: true,
-        shouldInitialize: didScreenTransitionEnd,
-    });
 
     const isEditing = assignCard?.isEditing;
 
-    const submit = (assignee: ListItem) => {
-        let nextStep: AssignCardStep = CONST.COMPANY_CARD.STEP.CARD;
-        const personalDetail = getPersonalDetailByEmail(assignee?.login ?? '');
-        const memberName = personalDetail?.firstName ? personalDetail.firstName : personalDetail?.login;
-        const data: Partial<AssignCardData> = {
-            email: assignee?.login ?? '',
-            cardName: getDefaultCardName(memberName),
-        };
+    const [selectedMember, setSelectedMember] = useState(assignCard?.data?.email ?? '');
+    const [searchTerm, debouncedSearchTerm, setSearchTerm] = useDebouncedState('');
+    const [shouldShowError, setShouldShowError] = useState(false);
 
+    const selectMember = (assignee: ListItem) => {
         Keyboard.dismiss();
-        if (assignee?.login === assignCard?.data?.email) {
+        setSelectedMember(assignee.login ?? '');
+        setShouldShowError(false);
+    };
+
+    const submit = () => {
+        let nextStep: AssignCardStep = CONST.COMPANY_CARD.STEP.CARD;
+        if (selectedMember === assignCard?.data?.email) {
             setAssignCardStepAndData({
                 currentStep: isEditing ? CONST.COMPANY_CARD.STEP.CONFIRMATION : nextStep,
                 isEditing: false,
@@ -89,17 +70,17 @@ function AssigneeStep({policy, feed}: AssigneeStepProps) {
             return;
         }
 
-        if (!policy?.employeeList?.[assignee?.login ?? '']) {
-            setAssignCardStepAndData({
-                currentStep: CONST.COMPANY_CARD.STEP.INVITE_NEW_MEMBER,
-                data: {
-                    email: assignee?.login ?? '',
-                    assigneeAccountID: assignee?.accountID ?? undefined,
-                },
-            });
-            setDraftInviteAccountID(assignee?.login ?? '', assignee?.accountID ?? undefined, policy?.id);
+        if (!selectedMember) {
+            setShouldShowError(true);
             return;
         }
+
+        const personalDetail = getPersonalDetailByEmail(selectedMember);
+        const memberName = personalDetail?.firstName ? personalDetail.firstName : personalDetail?.login;
+        const data: Partial<AssignCardData> = {
+            email: selectedMember,
+            cardName: getDefaultCardName(memberName),
+        };
 
         if (hasOnlyOneCardToAssign(filteredCardList)) {
             nextStep = CONST.COMPANY_CARD.STEP.TRANSACTION_START_DATE;
@@ -125,6 +106,9 @@ function AssigneeStep({policy, feed}: AssigneeStepProps) {
         Navigation.goBack();
     };
 
+    const shouldShowSearchInput = policy?.employeeList && Object.keys(policy.employeeList).length >= MINIMUM_MEMBER_TO_SHOW_SEARCH;
+    const textInputLabel = shouldShowSearchInput ? translate('workspace.card.issueNewCard.findMember') : undefined;
+
     const membersDetails = useMemo(() => {
         let membersList: ListItem[] = [];
         if (!policy?.employeeList) {
@@ -143,7 +127,7 @@ function AssigneeStep({policy, feed}: AssigneeStepProps) {
                 alternateText: email,
                 login: email,
                 accountID: personalDetail?.accountID,
-                isSelected: assignCard?.data?.email === email,
+                isSelected: selectedMember === email,
                 icons: [
                     {
                         source: personalDetail?.avatar ?? Expensicons.FallbackAvatar,
@@ -158,7 +142,7 @@ function AssigneeStep({policy, feed}: AssigneeStepProps) {
         membersList = sortAlphabetically(membersList, 'text', localeCompare);
 
         return membersList;
-    }, [isOffline, policy?.employeeList, assignCard?.data?.email, formatPhoneNumber, localeCompare]);
+    }, [isOffline, policy?.employeeList, selectedMember, formatPhoneNumber, localeCompare]);
 
     const sections = useMemo(() => {
         if (!debouncedSearchTerm) {
@@ -170,82 +154,23 @@ function AssigneeStep({policy, feed}: AssigneeStepProps) {
             ];
         }
 
-        const sectionsArr = [];
+        const searchValue = getSearchValueForPhoneOrEmail(debouncedSearchTerm, countryCode).toLowerCase();
+        const filteredOptions = tokenizedSearch(membersDetails, searchValue, (option) => [option.text ?? '', option.alternateText ?? '']);
 
-        if (!areOptionsInitialized) {
-            return [];
-        }
-
-        const searchValueForOptions = getSearchValueForPhoneOrEmail(debouncedSearchTerm, countryCode).toLowerCase();
-        const filteredOptions = tokenizedSearch(membersDetails, searchValueForOptions, (option) => [option.text ?? '', option.alternateText ?? '']);
-
-        sectionsArr.push({
-            title: undefined,
-            data: filteredOptions,
-            shouldShow: true,
-        });
-
-        // Selected options section
-        if (selectedOptionsForDisplay.length > 0) {
-            sectionsArr.push({
+        return [
+            {
                 title: undefined,
-                data: selectedOptionsForDisplay,
-            });
-        }
-
-        // Recent reports section
-        if (availableOptions.recentReports.length > 0) {
-            sectionsArr.push({
-                title: undefined,
-                data: availableOptions.recentReports,
-            });
-        }
-
-        // Contacts section
-        if (availableOptions.personalDetails.length > 0) {
-            sectionsArr.push({
-                title: undefined,
-                data: availableOptions.personalDetails,
-            });
-        }
-
-        // User to invite section
-        if (availableOptions.userToInvite) {
-            sectionsArr.push({
-                title: undefined,
-                data: [availableOptions.userToInvite],
-            });
-        }
-
-        return sectionsArr;
-    }, [
-        debouncedSearchTerm,
-        areOptionsInitialized,
-        countryCode,
-        membersDetails,
-        selectedOptionsForDisplay,
-        availableOptions.recentReports,
-        availableOptions.personalDetails,
-        availableOptions.userToInvite,
-    ]);
-
-    useEffect(() => {
-        searchInServer(searchTerm);
-    }, [searchTerm]);
+                data: filteredOptions,
+                shouldShow: true,
+            },
+        ];
+    }, [membersDetails, debouncedSearchTerm, countryCode]);
 
     const headerMessage = useMemo(() => {
-        const searchValue = searchTerm.trim().toLowerCase();
-        if (!availableOptions.userToInvite && CONST.EXPENSIFY_EMAILS_OBJECT[searchValue]) {
-            return translate('messages.errorMessageInvalidEmail');
-        }
-        return getHeaderMessage(
-            sections.some((section) => section.data.length > 0),
-            !!availableOptions.userToInvite,
-            searchValue,
-            countryCode,
-            false,
-        );
-    }, [searchTerm, availableOptions.userToInvite, sections, countryCode, translate]);
+        const searchValue = debouncedSearchTerm.trim().toLowerCase();
+
+        return getHeaderMessage(sections[0].data.length !== 0, false, searchValue, countryCode, false);
+    }, [debouncedSearchTerm, sections, countryCode]);
 
     return (
         <InteractiveStepWrapper
@@ -255,21 +180,29 @@ function AssigneeStep({policy, feed}: AssigneeStepProps) {
             stepNames={CONST.COMPANY_CARD.STEP_NAMES}
             headerTitle={translate('workspace.companyCards.assignCard')}
             enableEdgeToEdgeBottomSafeAreaPadding
-            onEntryTransitionEnd={() => setDidScreenTransitionEnd(true)}
         >
             <Text style={[styles.textHeadlineLineHeightXXL, styles.ph5, styles.mv3]}>{translate('workspace.companyCards.whoNeedsCardAssigned')}</Text>
             <SelectionList
-                textInputLabel={translate('selectionList.nameEmailOrPhoneNumber')}
+                textInputLabel={textInputLabel}
                 textInputValue={searchTerm}
                 onChangeText={setSearchTerm}
                 sections={sections}
                 headerMessage={headerMessage}
                 ListItem={UserListItem}
-                onSelectRow={submit}
+                onSelectRow={selectMember}
+                initiallyFocusedOptionKey={selectedMember}
                 shouldUpdateFocusedIndex
                 addBottomSafeAreaPadding
-                showLoadingPlaceholder={!areOptionsInitialized}
-                isLoadingNewOptions={!!isSearchingForReports}
+                footerContent={
+                    <FormAlertWithSubmitButton
+                        buttonText={translate(isEditing ? 'common.confirm' : 'common.next')}
+                        onSubmit={submit}
+                        isAlertVisible={shouldShowError}
+                        containerStyles={[!shouldShowError && styles.mt5]}
+                        addButtonBottomPadding={false}
+                        message={translate('common.error.pleaseSelectOne')}
+                    />
+                }
             />
         </InteractiveStepWrapper>
     );
