+++ src/types/onyx/AssignCard.ts
@@ -40,9 +40,6 @@ type AssignCardData = {
 
     /** Plaid accounts */
     plaidAccounts?: LinkAccount[] | PlaidAccount[];
-
-    /** The account ID of the cardholder */
-    assigneeAccountID?: number;
 };
 
 /** Model of assign card flow */
